Don't try to remove state dirs that don't exist

+++ src/distributed_rl.py
@@ -104,7 +104,7 @@ def check_resume(args: argparse.Namespace) -> None:
     else:
         resume = args.resume
 
-    if resume == "no":
+    if resume == "no" and resume_exists:
         shutil.rmtree(str(args.state_dir))
 
 def make_initial_filestructure(args: argparse.Namespace) -> None:
